Split out ExtendedArgumentParser into its own file
