+++ setup.py
@@ -4,7 +4,7 @@ import re
 
 from setuptools import setup, find_packages
 
-with open('buku/__init__.py', encoding='utf-8') as f:
+with open('buku/bukuconstants.py', encoding='utf-8') as f:
     version = re.search('__version__ = \'([^\']+)\'', f.read()).group(1)
 
 with open('README.md', encoding='utf-8') as f:
